deps/v8/test/inspector/debugger/step-into-break-on-async-call-expected.txt

Test for Debugger.stepInto with breakOnAsyncCall.

Running test: testSetTimeout
(anonymous) (test.js:0:0)
asyncCallStackTraceId is set
(anonymous) (test.js:0:17)
asyncCallStackTraceId is empty

Running test: testPromiseThen
(anonymous) (test.js:0:2)
asyncCallStackTraceId is set
(anonymous) (test.js:0:13)
asyncCallStackTraceId is empty
